Rename rarity module to count
