+++ README.md
@@ -8,9 +8,9 @@ ntStat is a toolkit that uses Bloom filters to track both *k*-mer count and dept
   - [C++ libraries](#c-libraries)
   - [Python packages](#python-packages)
 - [Modules](#modules)
+  - [`count`](#count)
   - [`filter`](#filter)
   - [`query`](#query)
-  - [`rarity`](#rarity)
   - [`hist`](#hist)
 - [Credits](#credits)
 - [License](#license)
@@ -59,9 +59,36 @@ tqdm
 
 # Modules
 
+## `count`
+
+Calculate TF-IDF information from sequencing data. This module outputs two counting Bloom filters: One for *k*-mer/seed counts (TF) and other for depths (DF). The number of reads will also be printed to stdout.
+
+```
+Usage: count [--help] [--version] [-k VAR] [-s VAR] -f VAR [-e VAR] [-b VAR] [--long] [-t VAR] -o VAR reads
+
+Positional arguments:
+  reads          path to sequencing data file(s) [nargs: 0 or more] [required]
+
+Optional arguments:
+  -h, --help     shows help message and exits 
+  -v, --version  prints version information and exits 
+  -k             k-mer length 
+  -s             path to spaced seeds file (one per line, if -k not specified) 
+  -f             path to k-mer spectrum file (from ntCard) [required]
+  -e             target output false positive rate [nargs=0..1] [default: 0.0001]
+  -b             output BF/CBF size (bytes) 
+  --long         optimize for long read data 
+  -t             number of threads [nargs=0..1] [default: 1]
+  -o             path to store output file [required]
+```
+
+- Set `-s` to count spaced seeds instead of *k*-mers. `-k` will be ignored in this case. The file pointed to by `-s` should be a plain text file containing spaced seed patterns, one per line, with `1` for care and `0` for don't care positions.
+- `-f` should point to the histogram file generated by [ntCard](https://github.com/bcgsc/ntCard). Use the same `-k` for ntCard and ntStat. For spaced seeds, set ntCard's `-k` parameter to the seed length.
+- Use `-b` to manually set the output data structure's size. Alternatively, set `-e` to the desired false positive rate (0.0001 by default) and ntStat will automatically calculate a suitable value for `-b`.
+
 ## `filter`
 
-Gathers *k*-mer counts and saves *k*-mers that occur at least `-cmin` times and at most `-cmax` times in a Bloom filter. Optionally, output can be a counting Bloom filter (set `--counts`), which includes the filtered *k*-mers' counts, resulting in a 8x larger output file.
+Gathers *k*-mer counts ONLY and saves *k*-mers that occur at least `-cmin` times and at most `-cmax` times in a Bloom filter. Optionally, output can be a counting Bloom filter (set `--counts`), which includes the filtered *k*-mers' counts, resulting in a 8x larger output file.
 
 ```
 Usage: ntstat filter [--help] [--version] [-k VAR] [-s VAR] -f VAR [-e VAR] [-b VAR] [-cmin VAR] [-cmax VAR] [--counts] [--long] [-t VAR] -o VAR reads
@@ -85,9 +112,7 @@ Optional arguments:
   -o             path to store output file [required]
 ```
 
-- Set `-s` to count spaced seeds instead of *k*-mers. `-k` will be ignored in this case. The file pointed to by `-s` should be a plain text file containing spaced seed patterns, one per line, with `1` for care and `0` for don't care positions.
-- `-f` should point to the histogram file generated by [ntCard](https://github.com/bcgsc/ntCard). Use the same `-k` for ntCard and ntStat. For spaced seeds, set ntCard's `-k` parameter to the seed length.
-- Use `-b` to manually set the output data structure's size. Alternatively, set `-e` to the desired false positive rate (0.0001 by default) and ntStat will automatically calculate a suitable value for `-b`.
+- Refer to the `count` module's notes for information about the parameters.
 
 ## `query`
 
@@ -108,32 +133,6 @@ options:
 
 - The input data can be any FASTA, FASTQ, or BAM. Compressed files are supported. To read from stdin, use `-`.
 - If the data structure contains spaced seed information, pass the path to the spaced seeds text file to `-s`.
-
-
-## `rarity`
-
-Calculate TF-IDF information from sequencing data. This module outputs two counting Bloom filters: One for *k*-mer/seed counts (TF) and other for depths (DF). The number of reads will also be printed to stdout.
-
-```
-Usage: rarity [--help] [--version] [-k VAR] [-s VAR] -f VAR [-e VAR] [-b VAR] [--long] [-t VAR] -o VAR reads
-
-Positional arguments:
-  reads          path to sequencing data file(s) [nargs: 0 or more] [required]
-
-Optional arguments:
-  -h, --help     shows help message and exits 
-  -v, --version  prints version information and exits 
-  -k             k-mer length 
-  -s             path to spaced seeds file (one per line, if -k not specified) 
-  -f             path to k-mer spectrum file (from ntCard) [required]
-  -e             target output false positive rate [nargs=0..1] [default: 0.0001]
-  -b             output BF/CBF size (bytes) 
-  --long         optimize for long read data 
-  -t             number of threads [nargs=0..1] [default: 1]
-  -o             path to store output file [required]
-```
-
-- Refer to the `filter` module's notes for information about the parameters.
 
 ## `hist`
